theRecommender/requirements.txt
Django==3.1.1
django-channels==0.7.0
asgi_redis
asgiref==3.1.2
channels-redis
django-channels-presence
pillow